+++ functions/public/config/Invoke-ActivateWindows.ps1
@@ -3,5 +3,5 @@ function Invoke-ActivateWindows {
         .DESCRIPTION
         Run Microsoft Activation Scripts (MAS) script 
     #>
-    Start-Process powershell -ArgumentList "-NoProfile -ExecutionPolicy Bypass -Command irm https://massgrave.dev/get | iex" -Verb RunAs
+    Start-Process powershell -ArgumentList "-NoProfile -ExecutionPolicy Bypass -Command irm https://get.activated.win | iex" -Verb RunAs
 }