+++ functions/public/config/Invoke-PanelAutologin.ps1
@@ -5,4 +5,5 @@ function Invoke-PanelAutologin {
     #>
     curl.exe -ss "https://live.sysinternals.com/Autologon.exe" -o $env:temp\autologin.exe # Official Microsoft recommendation https://learn.microsoft.com/en-us/sysinternals/downloads/autologon
     cmd /c $env:temp\autologin.exe
+    Write-Host "Autologin enabled" -ForegroundColor Green
 }